More warnings

+++ Source/cg_ControlGrisAudioProcessor.cpp
@@ -1277,7 +1277,7 @@ void ControlGrisAudioProcessor::processBlock([[maybe_unused]] juce::AudioBuffer<
         for (int i{}; i < mNumChannelsToAnalyse; ++i) {
             mDescriptorsBuffer.addFrom(0, 0, buffer, i, 0, buffer.getNumSamples());
         }
-        mDescriptorsBuffer.applyGain((1.0 / mNumChannelsToAnalyse) * mAudioAnalysisInputGainMultiplier);
+        mDescriptorsBuffer.applyGain((1.0f / mNumChannelsToAnalyse) * static_cast<float>(mAudioAnalysisInputGainMultiplier));
 
         auto bufferMagnitude = mDescriptorsBuffer.getMagnitude(0, mDescriptorsBuffer.getNumSamples());
         auto * channelData = mDescriptorsBuffer.getReadPointer(0);
