+++ Source/cg_LinkStrategies.cpp
@@ -588,7 +588,7 @@ SourceSnapshot FixedElevation::computeInitialStateFromFinalState_implementation(
 void LinearMin::computeParameters_implementation(Sources const & sources, SourcesSnapshots const & /*snapshots*/)
 {
     mBaseElevation = sources.getPrimarySource().getElevation();
-    mElevationPerSource = (ELEVATION_DIFF * mSourceLinkScale) / static_cast<float>((sources.size() - 1));
+    mElevationPerSource = (ELEVATION_DIFF * static_cast<float>(mSourceLinkScale)) / static_cast<float>((sources.size() - 1));
 }
 
 //==============================================================================
@@ -614,7 +614,7 @@ SourceSnapshot LinearMin::computeInitialStateFromFinalState_implementation([[may
 void LinearMax::computeParameters_implementation(Sources const & sources, SourcesSnapshots const & /*snapshots*/)
 {
     mBaseElevation = sources.getPrimarySource().getElevation();
-    mElevationPerSource = (ELEVATION_DIFF * mSourceLinkScale) / static_cast<float>((sources.size() - 1));
+    mElevationPerSource = (ELEVATION_DIFF * static_cast<float>(mSourceLinkScale)) / static_cast<float>((sources.size() - 1));
 }
 
 //==============================================================================
